Server/homelink/requirements.txt
annotated-types==0.7.0
anyio==4.4.0
asgiref==3.8.1
certifi==2024.6.2
colorama==0.4.6
distro==1.9.0
Django==4.2.11
django-cors-headers==4.3.1
django-environ==0.11.2
django-filter==24.3
djangorestframework==3.15.1
drf-yasg==1.21.8
exceptiongroup==1.2.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.5
httpx==0.27.0
idna==3.7
inflection==0.5.1
jiter==0.5.0
openai==1.40.0
packaging==24.1
psycopg2==2.9.9
pydantic==2.7.4
pydantic_core==2.18.4
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
sniffio==1.3.1
sqlparse==0.5.0
tqdm==4.66.4
typing_extensions==4.11.0
tzdata==2024.1
uritemplate==4.1.1
